+++ packages/EasyDocker/src/Console/EasyDockerApplication.php
@@ -11,7 +11,9 @@ final class EasyDockerApplication extends Application
 {
     use HelpfulApplicationTrait;
 
-    /** @var string */
+    /**
+     * @var string
+     */
     public const VERSION = '1.0.4';
 
     /**